feat: [US-006] - Admin- und Domain-Surfaces auf direkte shadcn-Nutzung vereinheitlichen
