--- src/components/dialog.tsx
@@ -1,111 +1,0 @@
-// SPDX-FileCopyrightText: 2026 SecPal
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-import * as DialogPrimitive from "@radix-ui/react-dialog";
-import { cn } from "@/lib/utils";
-
-const sizes = {
-  xs: "sm:max-w-xs",
-  sm: "sm:max-w-sm",
-  md: "sm:max-w-md",
-  lg: "sm:max-w-lg",
-  xl: "sm:max-w-xl",
-  "2xl": "sm:max-w-2xl",
-  "3xl": "sm:max-w-3xl",
-  "4xl": "sm:max-w-4xl",
-  "5xl": "sm:max-w-5xl",
-};
-
-export function Dialog({
-  open,
-  onClose,
-  size = "lg",
-  className,
-  children,
-}: {
-  open: boolean;
-  onClose: (open: false) => void;
-  size?: keyof typeof sizes;
-  className?: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <DialogPrimitive.Root
-      open={open}
-      onOpenChange={(nextOpen) => {
-        if (!nextOpen) {
-          onClose(false);
-        }
-      }}
-    >
-      <DialogPrimitive.Portal>
-        <DialogPrimitive.Overlay
-          data-slot="dialog-overlay"
-          className="fixed inset-0 z-40 bg-zinc-950/25 data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:animate-in data-[state=open]:fade-in-0 dark:bg-zinc-950/50"
-        />
-        <DialogPrimitive.Content
-          data-slot="dialog-content"
-          className={cn(
-            "fixed top-1/2 left-1/2 z-50 grid max-h-[calc(100dvh-2rem)] w-[calc(100%-2rem)] -translate-x-1/2 -translate-y-1/2 gap-4 overflow-x-hidden overflow-y-auto rounded-lg border border-zinc-200 bg-white p-6 text-zinc-950 shadow-lg duration-200 overscroll-contain data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[state=open]:animate-in data-[state=open]:fade-in-0 data-[state=open]:zoom-in-95 dark:border-zinc-800 dark:bg-zinc-950 dark:text-zinc-50",
-            sizes[size],
-            className
-          )}
-        >
-          {children}
-        </DialogPrimitive.Content>
-      </DialogPrimitive.Portal>
-    </DialogPrimitive.Root>
-  );
-}
-
-export function DialogTitle({
-  className,
-  ...props
-}: React.ComponentPropsWithoutRef<typeof DialogPrimitive.Title>) {
-  return (
-    <DialogPrimitive.Title
-      {...props}
-      className={cn(
-        "text-lg/6 font-semibold text-balance text-zinc-950 sm:text-base/6 dark:text-white",
-        className
-      )}
-    />
-  );
-}
-
-export function DialogDescription({
-  className,
-  ...props
-}: React.ComponentPropsWithoutRef<typeof DialogPrimitive.Description>) {
-  return (
-    <DialogPrimitive.Description
-      {...props}
-      className={cn(
-        "mt-2 text-pretty text-base/6 text-zinc-500 sm:text-sm/6 dark:text-zinc-400",
-        className
-      )}
-    />
-  );
-}
-
-export function DialogBody({
-  className,
-  ...props
-}: React.ComponentPropsWithoutRef<"div">) {
-  return <div {...props} className={cn("mt-6", className)} />;
-}
-
-export function DialogActions({
-  className,
-  ...props
-}: React.ComponentPropsWithoutRef<"div">) {
-  return (
-    <div
-      {...props}
-      className={cn(
-        "mt-8 flex flex-col-reverse items-center justify-end gap-3 *:w-full sm:flex-row sm:*:w-auto",
-        className
-      )}
-    />
-  );
-}
