+++ src/pages/Customers/CustomerDetail.tsx
@@ -30,12 +30,12 @@ import {
   DialogOverlay,
   DialogPortal,
   DialogTitle,
-  LinkButton,
-  PageLink,
-  PageText,
-  PageTitle,
-  StatusBadge,
-} from "../CustomerSites/ui";
+  CustomerSiteLinkButton as LinkButton,
+  CustomerSitePageLink as PageLink,
+  CustomerSitePageText as PageText,
+  CustomerSitePageTitle as PageTitle,
+  CustomerSiteStatusBadge as StatusBadge,
+} from "@/ui";
 import { useUserCapabilities } from "../../hooks/useUserCapabilities";
 import { isSafeMailtoTarget, isSafeTelTarget } from "../../utils/safeUrl";
