+++ src/pages/Employees/EmployeeList.tsx
@@ -17,26 +17,26 @@ import {
   Button,
   Card,
   CardContent,
-  DataTable,
+  EmployeeDataTable as DataTable,
   Field,
   FieldLabel,
   Input,
-  LinkButton,
-  PageText,
-  PageTitle,
+  EmployeeLinkButton as LinkButton,
+  EmployeePageText as PageText,
+  EmployeePageTitle as PageTitle,
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-  StatusBadge as EmployeeStatusBadge,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "./ui";
+  EmployeeStatusBadge,
+  EmployeeTable as Table,
+  EmployeeTableBody as TableBody,
+  EmployeeTableCell as TableCell,
+  EmployeeTableHead as TableHead,
+  EmployeeTableHeader as TableHeader,
+  EmployeeTableRow as TableRow,
+} from "@/ui";
 import { useUserCapabilities } from "../../hooks/useUserCapabilities";
 
 function StatusBadge({ status }: { status: EmployeeStatus }) {
