+++ src/ui/primitives.tsx
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import {
+  createContext,
   useCallback,
+  useContext,
   useId,
   useRef,
   useState,
@@ -37,9 +39,15 @@ import {
   Search,
   X,
 } from "lucide-react";
+import { PrefetchLink } from "@/components/PrefetchLink";
 import { getCspNonce } from "@/lib/cspNonce";
 import { cn } from "@/lib/utils";
-import { buttonVariants, type ButtonVariant, uiControlBase } from "./styles";
+import {
+  buttonVariants,
+  type ButtonVariant,
+  uiControlBase,
+  uiFocusRing,
+} from "./styles";
 
 export const Button = forwardRef(function Button(
   {
@@ -422,8 +430,7 @@ export function SearchableAutocompleteListbox({
   );
 }
 
-export interface SearchableAutocompleteOptionProps
-  extends ComponentPropsWithoutRef<"button"> {
+export interface SearchableAutocompleteOptionProps extends ComponentPropsWithoutRef<"button"> {
   highlighted?: boolean;
   slotPrefix: string;
 }
@@ -439,8 +446,9 @@ export function SearchableAutocompleteOption({
   void _type;
 
   const commandValue = typeof value === "string" ? value : undefined;
-  const commandItemProps =
-    props as ComponentPropsWithoutRef<typeof CommandItem>;
+  const commandItemProps = props as ComponentPropsWithoutRef<
+    typeof CommandItem
+  >;
 
   return (
     <CommandItem
@@ -637,9 +645,7 @@ export function SearchableCommandPopover({
                           ...(option.keywords ?? []),
                         ]}
                         disabled={option.disabled}
-                        data-current={
-                          option.value === value ? "" : undefined
-                        }
+                        data-current={option.value === value ? "" : undefined}
                         onSelect={selectOption}
                       >
                         <span className="min-w-0 flex-1 truncate">
@@ -1441,6 +1447,451 @@ export function DescriptionDetails({
       )}
       {...props}
     />
+  );
+}
+
+type CustomerSiteBadgeColor = "red" | "amber" | "lime" | "blue" | "zinc";
+
+const customerSiteBadgeColors = {
+  red: "bg-red-500/15 text-red-700 dark:bg-red-500/10 dark:text-red-400",
+  amber:
+    "bg-amber-400/20 text-amber-700 dark:bg-amber-400/10 dark:text-amber-400",
+  lime: "bg-lime-400/20 text-lime-700 dark:bg-lime-400/10 dark:text-lime-300",
+  blue: "bg-blue-500/15 text-blue-700 dark:bg-blue-500/10 dark:text-blue-400",
+  zinc: "bg-zinc-600/10 text-zinc-700 dark:bg-white/5 dark:text-zinc-400",
+} satisfies Record<CustomerSiteBadgeColor, string>;
+
+export function CustomerSitePageTitle({
+  level = 1,
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"h1"> & { level?: 1 | 2 }) {
+  const Component = level === 1 ? "h1" : "h2";
+
+  return (
+    <Component
+      data-slot="customer-site-heading"
+      className={cn(
+        level === 1
+          ? "text-2xl font-semibold tracking-normal text-zinc-950 dark:text-zinc-50"
+          : "text-base font-semibold tracking-normal text-zinc-950 dark:text-zinc-50",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function CustomerSitePageText({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"p">) {
+  return (
+    <p
+      data-slot="customer-site-text"
+      className={cn("text-sm text-zinc-600 dark:text-zinc-300", className)}
+      {...props}
+    />
+  );
+}
+
+export const CustomerSitePageLink = forwardRef<
+  HTMLAnchorElement,
+  ComponentPropsWithoutRef<typeof PrefetchLink>
+>(function CustomerSitePageLink({ className, ...props }, ref) {
+  return (
+    <PrefetchLink
+      ref={ref}
+      data-slot="customer-site-link"
+      className={cn(
+        "font-medium text-blue-600 underline-offset-4 hover:underline dark:text-blue-400",
+        uiFocusRing,
+        className
+      )}
+      {...props}
+    />
+  );
+});
+
+export const CustomerSiteLinkButton = forwardRef<
+  HTMLAnchorElement,
+  ComponentPropsWithoutRef<typeof PrefetchLink> & { variant?: ButtonVariant }
+>(function CustomerSiteLinkButton({ className, variant, ...props }, ref) {
+  return (
+    <PrefetchLink
+      ref={ref}
+      data-slot="customer-site-link-button"
+      className={cn(buttonVariants({ variant }), className)}
+      {...props}
+    />
+  );
+});
+
+export function CustomerSiteStatusBadge({
+  color = "zinc",
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"span"> & { color?: CustomerSiteBadgeColor }) {
+  return (
+    <Badge
+      data-slot="customer-site-status-badge"
+      className={cn(customerSiteBadgeColors[color], className)}
+      {...props}
+    />
+  );
+}
+
+export function CustomerSiteFormCheckboxField({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"div">) {
+  return (
+    <div
+      data-slot="customer-site-checkbox-field"
+      className={cn("flex items-center gap-3", className)}
+      {...props}
+    />
+  );
+}
+
+type EmployeeBadgeColor =
+  | "red"
+  | "orange"
+  | "amber"
+  | "yellow"
+  | "lime"
+  | "green"
+  | "rose"
+  | "sky"
+  | "zinc";
+
+const employeeBadgeColors = {
+  red: "bg-red-500/15 text-red-700 dark:bg-red-500/10 dark:text-red-400",
+  orange:
+    "bg-orange-500/15 text-orange-700 dark:bg-orange-500/10 dark:text-orange-400",
+  amber:
+    "bg-amber-400/20 text-amber-700 dark:bg-amber-400/10 dark:text-amber-400",
+  yellow:
+    "bg-yellow-400/20 text-yellow-700 dark:bg-yellow-400/10 dark:text-yellow-300",
+  lime: "bg-lime-400/20 text-lime-700 dark:bg-lime-400/10 dark:text-lime-300",
+  green:
+    "bg-green-500/15 text-green-700 dark:bg-green-500/10 dark:text-green-400",
+  rose: "bg-rose-400/15 text-rose-700 dark:bg-rose-400/10 dark:text-rose-400",
+  sky: "bg-sky-500/15 text-sky-700 dark:bg-sky-500/10 dark:text-sky-300",
+  zinc: "bg-zinc-600/10 text-zinc-700 dark:bg-white/5 dark:text-zinc-400",
+} satisfies Record<EmployeeBadgeColor, string>;
+
+export function EmployeeFieldset({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"fieldset">) {
+  return (
+    <fieldset
+      data-slot="employee-fieldset"
+      className={cn("space-y-6", className)}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeLegend({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"legend">) {
+  return (
+    <legend
+      data-slot="employee-legend"
+      className={cn(
+        "text-base font-semibold tracking-normal text-zinc-950 dark:text-zinc-50",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeAutocompleteListbox({
+  anchor,
+  open,
+  listboxId,
+  className,
+  children,
+}: {
+  anchor: ReactElement;
+  open: boolean;
+  listboxId: string;
+  className?: string;
+  children: ReactNode;
+}) {
+  return (
+    <SearchableAutocompleteListbox
+      anchor={anchor}
+      open={open}
+      listboxId={listboxId}
+      className={className}
+      slotPrefix="employee"
+    >
+      {children}
+    </SearchableAutocompleteListbox>
+  );
+}
+
+export function EmployeeAutocompleteOption({
+  className,
+  highlighted = false,
+  type = "button",
+  tabIndex = -1,
+  ...props
+}: ComponentPropsWithoutRef<"button"> & {
+  highlighted?: boolean;
+}) {
+  return (
+    <SearchableAutocompleteOption
+      className={className}
+      highlighted={highlighted}
+      slotPrefix="employee"
+      type={type}
+      tabIndex={tabIndex}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeCommandPopover({
+  label,
+  options,
+  value,
+  onValueChange,
+  placeholder,
+  searchPlaceholder,
+  emptyMessage,
+  disabled = false,
+  errorMessage,
+}: {
+  label: string;
+  options: CommandOption[];
+  value?: string;
+  onValueChange: (value: string) => void;
+  placeholder: string;
+  searchPlaceholder: string;
+  emptyMessage: string;
+  disabled?: boolean;
+  errorMessage?: string;
+}) {
+  return (
+    <SearchableCommandPopover
+      label={label}
+      options={options}
+      value={value}
+      onValueChange={onValueChange}
+      placeholder={placeholder}
+      searchPlaceholder={searchPlaceholder}
+      emptyMessage={emptyMessage}
+      disabled={disabled}
+      errorMessage={errorMessage}
+      slotPrefix="employee"
+    />
+  );
+}
+
+export function EmployeePageTitle({
+  level = 1,
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"h1"> & { level?: 1 | 2 | 3 }) {
+  const Component = level === 1 ? "h1" : level === 2 ? "h2" : "h3";
+
+  return (
+    <Component
+      data-slot="employee-heading"
+      className={cn(
+        level === 1
+          ? "text-2xl font-semibold tracking-normal text-zinc-950 dark:text-zinc-50"
+          : "text-base font-semibold tracking-normal text-zinc-950 dark:text-zinc-50",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function EmployeePageText({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"p">) {
+  return (
+    <p
+      data-slot="employee-text"
+      className={cn("text-sm text-zinc-600 dark:text-zinc-300", className)}
+      {...props}
+    />
+  );
+}
+
+export const EmployeePageLink = forwardRef<
+  HTMLAnchorElement,
+  ComponentPropsWithoutRef<typeof PrefetchLink>
+>(function EmployeePageLink({ className, ...props }, ref) {
+  return (
+    <PrefetchLink
+      ref={ref}
+      data-slot="employee-link"
+      className={cn(
+        "font-medium text-blue-600 underline-offset-4 hover:underline dark:text-blue-400",
+        uiFocusRing,
+        className
+      )}
+      {...props}
+    />
+  );
+});
+
+export const EmployeeLinkButton = forwardRef<
+  HTMLAnchorElement,
+  ComponentPropsWithoutRef<typeof PrefetchLink> & { variant?: ButtonVariant }
+>(function EmployeeLinkButton({ className, variant, ...props }, ref) {
+  return (
+    <PrefetchLink
+      ref={ref}
+      data-slot="employee-link-button"
+      className={cn(buttonVariants({ variant }), className)}
+      {...props}
+    />
+  );
+});
+
+export function EmployeeStatusBadge({
+  color = "zinc",
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"span"> & { color?: EmployeeBadgeColor }) {
+  return (
+    <Badge
+      data-slot="employee-status-badge"
+      className={cn(employeeBadgeColors[color], className)}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeDataTable({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"div">) {
+  return (
+    <div
+      data-slot="employee-table-shell"
+      className={cn(
+        "overflow-x-auto rounded-md border border-zinc-200 dark:border-zinc-800",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeTable({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"table">) {
+  return (
+    <table
+      data-slot="employee-table"
+      className={cn(
+        "min-w-full divide-y divide-zinc-200 text-left text-sm text-zinc-950 dark:divide-zinc-800 dark:text-zinc-50",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeTableHead(props: ComponentPropsWithoutRef<"thead">) {
+  return <thead data-slot="employee-table-head" {...props} />;
+}
+
+export function EmployeeTableBody({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"tbody">) {
+  return (
+    <tbody
+      data-slot="employee-table-body"
+      className={cn("divide-y divide-zinc-100 dark:divide-zinc-800", className)}
+      {...props}
+    />
+  );
+}
+
+const EmployeeTableRowLinkContext = createContext<{
+  to?: string;
+  title?: string;
+}>({});
+
+export function EmployeeTableRow({
+  className,
+  to,
+  title,
+  ...props
+}: ComponentPropsWithoutRef<"tr"> & { to?: string; title?: string }) {
+  return (
+    <EmployeeTableRowLinkContext.Provider value={{ to, title }}>
+      <tr
+        data-slot="employee-table-row"
+        className={cn(
+          "bg-white dark:bg-zinc-950",
+          to &&
+            "hover:bg-zinc-50 has-[[data-row-link]:focus-visible]:outline-2 has-[[data-row-link]:focus-visible]:-outline-offset-2 has-[[data-row-link]:focus-visible]:outline-blue-600 dark:hover:bg-zinc-900",
+          className
+        )}
+        {...props}
+      />
+    </EmployeeTableRowLinkContext.Provider>
+  );
+}
+
+export function EmployeeTableHeader({
+  className,
+  ...props
+}: ComponentPropsWithoutRef<"th">) {
+  return (
+    <th
+      data-slot="employee-table-header"
+      className={cn(
+        "px-4 py-3 text-xs font-medium uppercase tracking-normal text-zinc-500 dark:text-zinc-400",
+        className
+      )}
+      {...props}
+    />
+  );
+}
+
+export function EmployeeTableCell({
+  className,
+  children,
+  ...props
+}: ComponentPropsWithoutRef<"td">) {
+  const { to, title } = useContext(EmployeeTableRowLinkContext);
+  const [cellRef, setCellRef] = useState<HTMLTableCellElement | null>(null);
+
+  return (
+    <td
+      ref={to ? setCellRef : undefined}
+      data-slot="employee-table-cell"
+      className={cn("relative px-4 py-4 align-middle", className)}
+      {...props}
+    >
+      {to ? (
+        <PrefetchLink
+          data-row-link
+          to={to}
+          aria-label={title}
+          tabIndex={cellRef?.previousElementSibling === null ? 0 : -1}
+          className="absolute inset-0 focus-visible:outline-none"
+        />
+      ) : null}
+      {children}
+    </td>
   );
 }
